simforge-ai/backend/gateway/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
httpx==0.26.0
pydantic==2.6.1
jinja2==3.1.3
python-socketio==5.11.1
python-multipart==0.0.9
